consolidate rover data into single file
