--- app/constants/index.js
@@ -1,4 +1,0 @@
-export * from "./rover-cameras";
-export * from "./rover-image-paths";
-export * from "./rover-names";
-export * from "./search-queries";
