--- app/constants/rover-image-paths.js
@@ -1,6 +1,0 @@
-export const IMG_PATHS = {
-	curiosity: require("../../assets/img/curiosity-up-close.jpg"),
-	opportunity: require("../../assets/img/opportunity.jpg"),
-	perseverance: require("../../assets/img/perseverance.jpg"),
-	spirit: require("../../assets/img/spirit.jpg"),
-};
