--- app/constants/rover-names.js
@@ -1,6 +1,0 @@
-export const NAMES = {
-	curiosity: "Curiosity",
-	opportunity: "Opportunity",
-	perseverance: "Perseverance",
-	spirit: "Spirit",
-};
