--- app/constants/search-queries.js
@@ -1,1 +1,0 @@
-export const QUERIES = ["Martian SOL", "Earth Date", "Latest Photos"];
